Crate spl_elgamal_registry

Modules

instruction

processor

state

Constants

ID

The const program ID.

REGISTRY_ADDRESS_SEED

Seed for the ElGamal registry program-derived address

Functions

check_id

Returns true if given pubkey is the program ID.

get_elgamal_registry_address

Derives the ElGamal registry account address for the given wallet address

get_elgamal_registry_address_and_bump_seed

Derives the ElGamal registry account address and seed for the given wallet address

id

Returns the program ID.